# Some optimizations 