--- a/Cargo.toml
+++ b/Cargo.toml
@@ -8,6 +8,7 @@ members = [
 
 [profile.release]
 lto = true
+codegen-units = 1
 
 [workspace.dependencies]
 bpaf = "0.7.3"