--- a/frost/src/util/parsing.rs
+++ b/frost/src/util/parsing.rs
@@ -1,9 +1,11 @@
 use std::io::{self, Read};
 
+#[inline(always)]
 pub fn parse_u8(buf: &[u8]) -> io::Result<u8> {
     parse_u8_at(buf, 0)
 }
 
+#[inline(always)]
 pub fn parse_u8_at(buf: &[u8], index: usize) -> io::Result<u8> {
     let bytes = buf.get(index..index + 1).ok_or_else(|| {
         io::Error::new(
@@ -14,10 +16,12 @@ pub fn parse_u8_at(buf: &[u8], index: usize) -> io::Result<u8> {
     Ok(u8::from_le_bytes(bytes.try_into().unwrap()))
 }
 
+#[inline(always)]
 pub fn parse_le_u32(buf: &[u8]) -> io::Result<u32> {
     parse_le_u32_at(buf, 0)
 }
 
+#[inline(always)]
 pub fn parse_le_u32_at(buf: &[u8], index: usize) -> io::Result<u32> {
     let bytes = buf.get(index..index + 4).ok_or_else(|| {
         io::Error::new(
@@ -28,10 +32,12 @@ pub fn parse_le_u32_at(buf: &[u8], index: usize) -> io::Result<u32> {
     Ok(u32::from_le_bytes(bytes.try_into().unwrap()))
 }
 
+#[inline(always)]
 pub fn parse_le_u64(buf: &[u8]) -> io::Result<u64> {
     parse_le_u64_at(buf, 0)
 }
 
+#[inline(always)]
 pub fn parse_le_u64_at(buf: &[u8], index: usize) -> io::Result<u64> {
     let bytes = buf.get(index..index + 8).ok_or_else(|| {
         io::Error::new(
@@ -42,6 +48,7 @@ pub fn parse_le_u64_at(buf: &[u8], index: usize) -> io::Result<u64> {
     Ok(u64::from_le_bytes(bytes.try_into().unwrap()))
 }
 
+#[inline(always)]
 pub fn get_lengthed_bytes(reader: &mut impl Read) -> io::Result<Vec<u8>> {
     // Get a vector of bytes from a reader when the first 4 bytes are the length
     // Ex: with <header_len><header> or <data_len><data>, this function returns either header or data