This PR reduces codegen units to 1, and adds some inlines. Overall, it's about close to a 20% speedup. Overall binary size is reduced. 

On a linux desktop:
```bash
frost$ hyperfine --warmup 3 "./target/release/frost_orig info ./frost/tests/fixtures/test_large.bag" "./target/release/frost_cg info ./frost/tests/fixtures/test_large.bag" "./target/release/frost info ./frost/tests/fixtures/test_large.bag"
Benchmark 1: ./target/release/frost_orig info ./frost/tests/fixtures/test_large.bag
  Time (mean ± σ):     658.6 ms ±   1.0 ms    [User: 421.1 ms, System: 237.5 ms]
  Range (min … max):   657.1 ms … 659.7 ms    10 runs
 
Benchmark 2: ./target/release/frost_cg info ./frost/tests/fixtures/test_large.bag
  Time (mean ± σ):     575.9 ms ±   2.6 ms    [User: 334.6 ms, System: 241.3 ms]
  Range (min … max):   573.7 ms … 582.7 ms    10 runs
 
Benchmark 3: ./target/release/frost info ./frost/tests/fixtures/test_large.bag
  Time (mean ± σ):     555.9 ms ±   2.4 ms    [User: 322.2 ms, System: 233.6 ms]
  Range (min … max):   552.8 ms … 558.5 ms    10 runs
 
Summary
  './target/release/frost info ./frost/tests/fixtures/test_large.bag' ran
    1.04 ± 0.01 times faster than './target/release/frost_cg info ./frost/tests/fixtures/test_large.bag'
    1.18 ± 0.01 times faster than './target/release/frost_orig info ./frost/tests/fixtures/test_large.bag'
```

On a macbook:
```bash
Benchmark 1: ./target/release/frost_orig info ./frost/tests/fixtures/decompressed_large.bag
  Time (mean ± σ):      1.212 s ±  0.051 s    [User: 0.753 s, System: 0.456 s]
  Range (min … max):    1.144 s …  1.281 s    10 runs

Benchmark 2: ./target/release/frost_cg info ./frost/tests/fixtures/decompressed_large.bag
  Time (mean ± σ):      1.016 s ±  0.042 s    [User: 0.578 s, System: 0.435 s]
  Range (min … max):    0.969 s …  1.092 s    10 runs

Benchmark 3: ./target/release/frost info ./frost/tests/fixtures/decompressed_large.bag
  Time (mean ± σ):      1.001 s ±  0.056 s    [User: 0.569 s, System: 0.430 s]
  Range (min … max):    0.899 s …  1.107 s    10 runs

Summary
  './target/release/frost info ./frost/tests/fixtures/decompressed_large.bag' ran
    1.01 ± 0.07 times faster than './target/release/frost_cg info ./frost/tests/fixtures/decompressed_large.bag'
    1.21 ± 0.08 times faster than './target/release/frost_orig info ./frost/tests/fixtures/decompressed_large.bag'
```